third_party/WebKit/LayoutTests/fast/mediastream/RTCPeerConnection-localDescription-expected.txt

Tests RTCPeerConnection localDescription.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS pc.setLocalDescription(null) threw exception TypeError: Failed to execute 'setLocalDescription' on 'RTCPeerConnection': parameter 1 is not of type 'RTCSessionDescription'..
PASS pc.setLocalDescription(sessionDescription, requestSucceeded1, requestFailed1); did not throw exception.
PASS requestSucceeded was called.
PASS pc.setLocalDescription(sessionDescription, requestSucceeded2, requestFailed2); did not throw exception.
PASS requestFailed was called.
PASS pc.localDescription.type is "offer"
PASS pc.localDescription.sdp is "local"
PASS pc.localDescription.type is "offer"
PASS pc.localDescription.sdp is "local"
PASS successfullyParsed is true

TEST COMPLETE
